### 2021-09-21 #3745 @whedon
[PRE REVIEW]: cofad: An R package and shiny app for contrast analysis
**Submitting author:** @johannes-titz (<a href="http://orcid.org/0000-0002-1102-5719">Johannes Titz</a>)
**Repository:** <a href="https://github.com/johannes-titz/cofad" target ="_blank">https://github.com/johannes-titz/cofad</a>
**Version:** v0.2.0
