+++ paper/library_paper.bib
@@ -1,3 +1,15 @@
+@article{furr2004,
+  title = {Interpreting Effect Sizes in Contrast Analysis},
+  author = {Furr, R. Michael},
+  date = {2004},
+  journaltitle = {Understanding Statistics},
+  volume = {3},
+  pages = {1--25},
+  publisher = {{Routledge}},
+  doi = {10.1207/s15328031us0301_1},
+  file = {/home/jt/Zotero/storage/GZXAIE9Y/Furr - TEACHING ARTICLES Interpreting Effect Sizes in Con.pdf}
+}
+
 @book{rosenthal1985,
   title = {Contrast Analysis: {{Focused}} Comparisons in the Analysis of Variance},
   author = {Rosenthal, Robert and Rosnow, Ralph L},
