### 2021-11-01 #3822 @chartgerink
@johannes-titz could you please update the short DOIs to full length DOIs? Specifically these two

```
- 10/gdz7q2 is OK
- 10/b5f2gc is OK
```
