### 2021-11-01 1050f46
update short DOIs to full length DOIs
