+++ paper/library_paper.bib
@@ -11,7 +11,7 @@
 }
 
 @book{rosenthal1985,
-  title = {Contrast Analysis: {{Focused}} Comparisons in the Analysis of Variance},
+  title = {Contrast Analysis: Focused Comparisons in the Analysis of Variance},
   author = {Rosenthal, Robert and Rosnow, Ralph L},
   date = {1985},
   publisher = {{Cambridge University Press}},
@@ -19,7 +19,7 @@
 }
 
 @book{Rosenthal2000,
-  title = {Contrasts and Effect Sizes in Behavioral Research: {{A}} Correlational Approach},
+  title = {Contrasts and Effect Sizes in Behavioral Research: A Correlational Approach},
   author = {Rosenthal, Robert and Rosnow, Ralph L and Rubin, Donald B},
   date = {1999},
   publisher = {{Cambridge University Press}},
@@ -31,12 +31,12 @@
   title = {Forschungsmethoden und Statistik f\"ur Psychologen und Sozialwissenschaftler},
   author = {Sedlmeier, Peter and Renkewitz, Frank},
   date = {2018},
-  edition = {3rd ed.},
+  edition = {3},
   publisher = {{Pearson Studium}},
   location = {{Hallbergmoos, Germany}},
   abstract = {Das beliebte und h\"aufig eingesetzte Lehrbuch bietet eine umfassende und verst\"andliche Einf\"uhrung in alle wichtigen Grundlagen der psychologischen und sozialwissenschaftlichen Forschungsmethodik. Auf die Bed\"urfnisse von Studenten angepasst, werden Konzepte und Verfahren m\"oglichst intuitiv und anhand zahlreicher einpr\"agsamer Beispiele aus dem Forschungsalltag erkl\"art ohne dabei an Pr\"azision einzub\"u\ss en. Eine klare Sprache und viele einpr\"agsame Beispiele unterst\"utzen den Lernprozess..In der Neuauflage wurden diverse neuere Entwicklungen mit aufgenommen. So wurden u.a. der Mehrebenenanalyse, Bootstrap-Verfahren und Randomisierungstests, ~dem Bayesianischen Ansatz sowie dem Kritischen Rationalismus als derzeit wohl wichtigstem wissenschaftstheoretischen Ansatz deutlich mehr Platz einger\"aumt. Neu hinzugekommen sind zudem Kapitel zu~Problemen der klassischen Inferenzstatistik in der Forschungspraxis, zu~Replikation, Pr\"aregistrierung, Open Science~sowie zur~Experimentellen Einzelfallanalyse.Teil I Grundlagen und Konzepte1.~~~~ Alltagswissen versus Wissenschaft: Beispiel Psychologie2.~~~~ Wissenschaftstheorie, Theorien und Hypothesen3.~~~~ Messen und Testen4.~~~~ Datenerhebung: Befragung und Beobachtung5.~~~~ Experimentelle DesignsTeil II Deskriptive und explorative Datenanalyse6.~~~~ Lage- und Streuungsma\ss e7.~~~~ Korrelation8.~~~~ Lineare Regression9.~~~~ Effektgr\"o\ss enTeil III Inferenzstatistik10.~~ Grundlagen der Inferenzstatistik11.~~ Konfidenzintervalle12.~~ Signifikanztests13.~~ t-Tests14.~~ Der F-Test in der einfaktoriellen Varianzanalyse15.~~ Weitere F-Tests16.~~ Kontrastanalyse17.~~ Verfahren zur Analyse nominalskalierter Daten: Chi Quadrat-Tests18.~~ Verfahren zur Analyse ordinalskalierter Daten19.~~ Resampling-
[…]
