### 2021-11-01 #3822 @johannes-titz
@chartgerink Thanks for the update! I changed the short DOIs to full length DOIs. Is this OK?
