### 2021-11-08 #3822 @svmiller
I think what Titz and Burkhardt offer here merits publication in the Journal of Open Source Software. The authors advocate for greater use of a contrast analysis for testing hypotheses about means in a factorial design. The authors note that alternatives to the contrast analysis exist, though with important limitations. To better advocate its use, the authors offer a Shiny app that allows users to upload their own data to do their own contrast analyses.

I think the authors have provided an important contribution to researchers working in psychology and the social sciences (who would assuredly benefit from these methods). I was able to upload a pilot data set and calculate my own contrast analyses, so the software behaves as it should. I'm also a big proponent of rewarding those who make software to make all our lives easier. I recommend JOSS proceed with publication here.
