### 2021-11-15 #3822 @chartgerink
I haven't heard back from @tianxzhu, but given the positive review from @svmiller and no negative comments from @tianxzhu - I will move forward and am happy to accept this for publication, @johannes-titz 😊

@johannes-titz could you please deposit the software in Zenodo/FigShare and share the DOI with me? I will initiate some of whedon's checks in the meantime.
