+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: cofad
 Type: Package
 Title: Contrast Analyses for Factorial Designs
-Version: 0.2.0
+Version: 0.2.1
 Authors@R: c(person("Markus", "Burkhardt",
   email = "markus.burkhardt@psychologie.tu-chemnitz.de",
   role = c("aut", "cre")),
