### 2021-11-15 #3822 @johannes-titz
@chartgerink Thank you very much! I just made a final commit for the release and released version 0.2.1 on github (tag v0.2.1), which is also linked at zenodo (https://zenodo.org/record/5702779) with the following doi: 10.5281/zenodo.5702779

Do you need anything else?

@svmiller Thank you for such a positive feedback! I hope the software will be as useful as your review suggests :).

@tianxzhu Thank you very much for your time and effort to review the submission!
